feat: finish implementing `urfave/cli` and extra methods
